Create cartSlice using Redux Toolkit
I created the cartSlice and created the ability to add and delete items from the cart, increase and decrease the quantity of the items in the cart and also the ability to clear the whole cart.

I also created the reusable DeleteItem and UpdateItemQuantity components as I will be using them in the Menu as well as the Cart.

I also connected the Cart Overview with the data and only make the overview appear when there is something in the cart.