Add ability to create and search orders
I created a react router action which takes the data from the form and converts it into an object. I created a hidden input within the form that contains the cart data. That new object is then sent to a new async function I created which then sends the data to the API. The user is then redirected to the order page which contains all the information from the cart and form.

I also added the ability to search for any existing order within the search input in the header.

I added an Error component that will display when there is an error.